inside-secure: add new GPLv2 "mini" firmware for the EIP197 driver
This adds the "minifw" version of the EIP197 firmware, which the inside-
secure driver will use as a fallback if the original full-featured
firmware cannot be found. This allows for using the inside-secure driver
and hardware without access to "official" firmware only available under
NDA. Note that this "minifw" was written by me (Pascal) specifically for
this driver and I am allowed by my employer, Verimatrix, to release this
under the GPLv2 licence.

Signed-off-by: Pascal van Leeuwen <pvanleeuwen@verimatrix.com>